requirements_arm.txt
# Core dependencies for Windows on ARM
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
shiboken6==6.10.2

# Data Processing (ARM64 native wheels usually available)
numpy>=2.0.0
pandas>=2.2.0
python-docx==1.1.2
openpyxl==3.1.5

# AI & Runtime (Natively optimized for ARM/NPU)
onnxruntime-genai==0.11.4
# onnxruntime-qnn # Install manually if using Qualcomm NPU: pip install onnxruntime-qnn
onnxruntime-directml>=1.19.0 # Fallback for GPU acceleration on Windows

# Connectivity & Utilities
requests==2.32.3
psutil==6.1.1
python-dateutil==2.9.0
pyyaml==6.0.2
loguru==0.7.2
tqdm==4.67.1
pydantic==2.10.5
huggingface_hub>=0.24.0

# Optional/Fallback AI
ollama==0.6.1

# Performance & Hardware
coloredlogs==15.0.1
humanfriendly==10.0
rich==14.3.1
